eos_project/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
alembic
sqladmin[full]
passlib[bcrypt]
python-jose[cryptography]
